Get the pulsing animation done
