+++ Assets/02_Scripts/Animations.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 5fd3aaa333414428bc18e8684a6e0bbd
+timeCreated: 1722433124
